Generalize check of the third party RPMs
This is initial draft to generalize current check of installed
third party RPMs that are not part of the installed system
distribution.

Original actor has been written only for RHEL systems and expected
that the only vendor as such is Red Hat. However, in case of other
distributions it's not true. So such a report could be confusing for
users and could set wrong expectations.

List of changes:
 * Rename and move redhatsignedrpmcheck actor to <NEW_NAME>
 * Update docstrings
 * Update the report, respecting name of the installed system
 * ...
